Changed types to int32_t, int64_t and the like

+++ blib/RGReads.h
@@ -8,16 +8,16 @@
 
 void RGReadsFindMatches(RGIndex*, RGBinary*, RGMatch*, char*, int*, int, int, int, int, int, int);
 void RGReadsGenerateReads(char*, RGIndex*, RGReads*, char, int*, int, int, int, int, int, int);
-void RGReadsGenerateMismatches(char*, int, char, int, unsigned int, unsigned int*, unsigned int*, unsigned int, int, RGReads*);
-void RGReadsGenerateMismatchesHelper(char*, int, char, int, unsigned int, unsigned int*, unsigned int*, unsigned int, int, int, RGReads*, char*, int, int);
-void RGReadsGenerateDeletions(char*, int, char, int, unsigned int, unsigned int*, unsigned int*, unsigned int, int, RGReads*);
-void RGReadsGenerateDeletionsHelper(char*, int, char, int, unsigned int, unsigned int*, unsigned int*, unsigned int, int, int, int, int, RGReads*, char*, int, int);
-void RGReadsGenerateInsertions(char*, int, char, int, unsigned int, unsigned int*, unsigned int*, unsigned int, int, RGReads*);
-void RGReadsGenerateInsertionsHelper(char*, int, char, int, unsigned int, unsigned int*, unsigned int*, unsigned int, int, int, int, int, RGReads*, char*, int, int);
-void RGReadsGenerateGapDeletions(char*, int, char, int, unsigned int, unsigned int*, unsigned int*, unsigned int, int, RGReads*);
-void RGReadsGenerateGapDeletionsHelper(char*, int, char, int, unsigned int, unsigned int*, unsigned int*, unsigned int, int, RGReads*, char*);
-void RGReadsGenerateGapInsertions(char*, int, char, int, unsigned int, unsigned int*, unsigned int*, unsigned int, int, RGReads*);
-void RGReadsGenerateGapInsertionsHelper(char*, int, char, int, unsigned int, unsigned int*, unsigned int*, unsigned int, int, RGReads*, char*);
+void RGReadsGenerateMismatches(char*, int, char, int, int32_t, int32_t*, int32_t*, int64_t, int, RGReads*);
+void RGReadsGenerateMismatchesHelper(char*, int, char, int, int32_t, int32_t*, int32_t*, int64_t, int, int, RGReads*, char*, int, int);
+void RGReadsGenerateDeletions(char*, int, char, int, int32_t, int32_t*, int32_t*, int64_t, int, RGReads*);
+void RGReadsGenerateDeletionsHelper(char*, int, char, int, int32_t, int32_t*, int32_t*, int64_t, int, int, int, int, RGReads*, char*, int, int);
+void RGReadsGenerateInsertions(char*, int, char, int, int32_t, int32_t*, int32_t*, int64_t, int, RGReads*);
+void RGReadsGenerateInsertionsHelper(char*, int, char, int, int32_t, int32_t*, int32_t*, int64_t, int, int, int, int, RGReads*, char*, int, int);
+void RGReadsGenerateGapDeletions(char*, int, char, int, int32_t, int32_t*, int32_t*, int64_t, int, RGReads*);
+void RGReadsGenerateGapDeletionsHelper(char*, int, char, int, int32_t, int32_t*, int32_t*, int64_t, int, RGReads*, char*);
+void RGReadsGenerateGapInsertions(char*, int, char, int, int32_t, int32_t*, int32_t*, int64_t, int, RGReads*);
+void RGReadsGenerateGapInsertionsHelper(char*, int, char, int, int32_t, int32_t*, int32_t*, int64_t, int, RGReads*, char*);
 
 void RGReadsRemoveDuplicates(RGReads*);
 void RGReadsQuickSort(RGReads*, int, int);
